requirements.txt
certifi==2018.10.15
chardet==3.0.4
configparser==3.5.0
idna==2.7
requests==2.19.1
urllib3==1.23
wordnik-py3==2.1.2
